Add files via upload
rodando mais ou menos

+++ mainF.c
@@ -12,7 +12,7 @@
 
 #include "fila.h"
 
-#define MAX_ARGS  10	
+#define MAX_ARGS  10
 #define MAX_STRING 80
 #define TRUE 1
 #define NFILAS 1
@@ -23,6 +23,7 @@
 ptFila filaIO;
 ptFila f1,f2,f3;
 ptFila filas[3];
+
 
 ptFila fAtual();
 ptFila fReset();
@@ -58,10 +59,11 @@ int main(int argc, char **argv){
 	// associa a memória compartilhada ao processo
 	flag = (int*) shmat (segmento, 0, 0);
 
-	f1 = FILA_cria(2);
-	//f1 = FILA_cria(1);
-	//f2 = FILA_cria(2);
-	//f3 = FILA_cria(4);
+	
+	f1 = FILA_cria(1);
+	f2 = FILA_cria(2);
+	f3 = FILA_cria(4);
+	
 	filaIO = FILA_cria(666);
 
 	signal(SIGUSR1, processoIO);
@@ -74,11 +76,11 @@ int main(int argc, char **argv){
 		read(0, comando, 81);
 		//printf("comando lido inicio:%s\n", comando);
 		if(strcmp(comando,"done")==0) break;
-		criaNovoProcesso(f1, comando);
+		criaNovoProcesso(f3, comando);
 		*flag-=1;			
 	}
-
-	id = FILA_topId(fAtual());
+	indexAtual =2;
+	id = FILA_topId(f3);
 	printf("Iniciando processo %d\n", id);
 	kill(id, SIGCONT);
 
@@ -99,56 +101,69 @@ int main(int argc, char **argv){
 		if(!FILA_vazia(filaIO))
 			FILA_atualizaIO(filaIO, tempo);
 				
+		
+
+		//printf("passou 1 u.t. agora estamos em %d\n", tempo);
 		id = FILA_topId(fAtual());
+		printf("id= %d \n",id);
 		if(id<=0){
 			printf("fila atual vazia, procurando outro processo\n");
 			if(fReset() == NULL  && FILA_vazia(filaIO)){
-				printf("todas as filas estao vazias\n");
+				printf("todas as filas estao vazia\n");
 				//encerra(0);
 			}
-			else if(!FILA_vazia(filaIO))
-				printf("processo aguardando I/O\n");
+			else if(!FILA_vazia(filaIO)){
+				printf("processo aguardando I/O");
+				continue;
+				}
 			else
-				printf("processo ocioso\n");	
-		}		
-
+				printf("processo ocioso");
+		
+		}
+printf("A1\n");
 		tempoRestante = FILA_tempoRestante(fAtual(), tempo);
 		printf("tempoRestante:%d do processo %d\n", tempoRestante, id);
-		
-		/* Se o tempo acabou, ou o programa solicitou I/O, então algum processo deve ser trocado */
-		if(solicitouIO || tempoRestante<0){
-
-			if(solicitouIO)
-				tempoRestante = FILA_tempoRestante(fAtual(), FILA_getTempoIO(fAtual()));
+	
+		/* Se o tempo acabou, ou o programa solicitou I/O, então algum processo deve trocar */
+		if(solicitouIO && tempoRestante<0){
 
 			if(id>0)
 				kill(id, SIGSTOP);
-			else
+			else{
 				printf("BUG NOUTRO LUGAR\n");
-
+			}
+printf("A2\n");
 			/* Caso um processo tenha extrapolado o tempo, ele deve ser movido para uma fila de nivel mais alto e menor prioridade */
 			if(!solicitouIO && tempoRestante<0 ){
 				printf("colocando processo %d numa fila mais alta\n", id);	
 				FILA_remove(fAtual());
 				FILA_insere(fProx(), id, tempo);//coloca numa mais alto	
 			}
-			/* Caso um processo tenha terminado antes do tempo esperado, ele deve ser movido para uma fila de nivel mais baixo e maior prioridade.
-			Temos que colocar o processo que pediu I/O em uma fila de I/O */
-			else if(solicitouIO && tempoRestante >0){
-				printf("colocando processo %d numa fila mais baixa\n", id);	
-				printf("colocando processo %d em I/O\n", id);
-				FILA_comecaIO(fAtual(), fAnt(), filaIO, tempo);//coloca numa mais baixo	
+		
+
+			id = FILA_topId(fAtual());
+			solicitouIO = 0;
+printf("A3\n");
+			if(id==-1){
+				printf("fila vazia\n");
+				if(FILA_vazia(filaIO)){
+					printf("Nenhum processo restante, fechando o programa");
+					encerra(0);
+				}
 			}
-			/*	Dentro da faixa de 'acabou no tempo certo'
-				Caso o processo tenha terminado no tempo correto, ele é movido para o final da fila 
-			Temos que colocar o processo que pediu I/O em uma fila de I/O */	
-			else if(solicitouIO && tempoRestante<=0){
-				printf("colocando processo %d no final da mesma fila\n", id);
-				printf("colocando processo %d em I/O\n", id);			
-				FILA_comecaIO(fAtual(), fAtual(), filaIO, tempo);//coloca na mesma
+			else{
+				printf("executando processo %d, que e o proximo da fila\n", id);
+				FILA_comecaCPU(fAtual(), tempo);
+
+				if(id>0)
+					kill(id, SIGCONT);
+				else{
+					printf("BUG ACOLA\n");
+				}
 			}
+
 		}
-		solicitouIO = 0;
+printf("A4\n");
 		/*
 			Agora vamos checar se existe algum outro processo numa fila de priorade maior que está esperando para ser executado.
 			Se houver, precisamos trocar da fila atual para a fila dele.
@@ -160,12 +175,10 @@ int main(int argc, char **argv){
 			printf("pula para a fila de cima\n");
 			/* Precisamos parar o processo que estava rodando antes */
 			id = FILA_topId(filaTemp);
-			if(id>0)
-				kill(id, SIGSTOP);
+			kill(id, SIGSTOP);
 			/* E continuar o processo atual */
 			id = FILA_topId(fAtual());
-			if(id>0)
-				kill(id, SIGCONT);
+			kill(id, SIGCONT);
 		}
 
 
@@ -194,11 +207,11 @@ void criaNovoProcesso(ptFila f, char *comando){
 	}
 	id = fork();
 	if(id==0){
-		printf("criado:%d por: %d - tempo %d - Novo processo\n%s\n", getpid(), getppid(), tempo, comando);
+		printf("criado:%d por: %d - tempo %d - Novo processo id\n%s\n", getpid(), getppid(), tempo, comando);
 		execv("./programa", args);
 		printf("dafuq");
 		encerra(1);
-	}
+	}	
 	FILA_insere(f,id, 0);
 	printf("getpid %d parando processo %d\n", getpid(), id);
 	if(id>0)
@@ -210,10 +223,33 @@ void criaNovoProcesso(ptFila f, char *comando){
 }
 
 void processoIO(int sinal){
+	int tempoRestante;
 	int id = FILA_topId(fAtual());
 	printf("Processo %d solicitou IO em tempo %d\n", id, tempo);
-	solicitouIO = 1;
-	FILA_setTempoIO(fAtual(), tempo);
+	kill(id, SIGSTOP);
+		
+	tempoRestante = FILA_tempoRestante(fAtual(), tempo);
+	/* Caso um processo tenha terminado antes do tempo esperado, ele deve ser movido para uma fila de nivel mais baixo e maior prioridade.
+		Temos que colocar o processo que pediu I/O em uma fila de I/O */
+		/*else*/ if(tempoRestante >0){
+			printf("colocando processo %d numa fila mais baixa\n", id);	
+			printf("colocando processo %d em I/O\n", id);
+			FILA_comecaIO(fAtual(), fAnt(), filaIO, tempo);	//coloca numa mais baixo		
+			
+		}
+		/*	Dentro da faixa de 'acabou no tempo certo'
+			Caso o processo tenha terminado no tempo correto, ele é movido para o final da fila 
+		Temos que colocar o processo que pediu I/O em uma fila de I/O */	
+		else if(tempoRestante<=0){
+			printf("colocando processo %d no final da mesma fila\n", id);
+			printf("colocando processo %d em I/O\n", id);			
+			FILA_comecaIO(fAtual(), fAtual(), filaIO, tempo);//coloca na mesma
+	
+		}
+
+
+	if(id>0)
+		solicitouIO = 1;
 }
 
 void processoTermina(int sinal){
@@ -237,51 +273,144 @@ void processoTermina(int sinal){
 /* Funcoes de controle de fila */
 
 ptFila fAtual(){
-	/*
-	if(indexAtual<0 || indexAtual>2) return NULL;
-	return filas[indexAtual];
-	*/
-	return f1;
+	int i;
+	printf("Index atual %d\n",indexAtual);
+//	if(indexAtual<0 || indexAtual>2) return NULL;
+//	return filas[indexAtual];
+
+	switch (indexAtual)
+{
+   case 0:
+     return f1;
+   break;
+   
+   case 1:
+     return f2;
+   break;
+      
+   case 2:
+     return f3;
+   break;
+   default: //erro???
+     return NULL;
 }
+//	return filas[i];*/
+}	
 
 /* 
 	Retorna primeira fila nao vazia, atribui index a indexAtual.
 	Retorna NULL se nenhuma fila estiver vazia, atribui -1 a indexAtual.
 */
 ptFila fReset(){
-	/*
+	
 	int i;
-	indexAtual = 0;
-	for(i=0;i<NFILAS;i++){
-		if(!FILA_vazia(filas[i])){
-			indexAtual = i;
-			return filas[indexAtual];
-		}
+	printf("RA\n");
+	if(filas[0]==NULL || filas[1] == NULL)
+		printf("filas nulas");
+//	indexAtual = 0;
+//	for(i=0;i<NFILAS;i++){
+//		if(!FILA_vazia(filas[i])){
+//		indexAtual = i;			
+//		}
+//	}
+	if (!FILA_vazia(f1))
+	{
+		indexAtual = 0;
+		return f1;
 	}
+	if (!FILA_vazia(f2))
+	{
+		indexAtual = 1;
+		return f2;
+	}
+	if (!FILA_vazia(f3))
+	{
+		indexAtual = 2;
+		return f3;
+	}	
+
+	printf("RF\n");
+	switch (indexAtual)
+	{
+	   case 0:	   	
+		 return f1;
+	   
+	   
+	   case 1:
+		 return f2;
+	   
+		  
+	   case 2:
+		 return f3;
+	   
+	   default: //erro???
+	   	indexAtual = -1;
+		 return NULL;
+	}
+	
 	indexAtual = -1;
+	
 	return NULL;
-	*/
-	return f1;
-}
+	}
 
 /* Retorna fila, um nivel de fila maior do que a atual (note que o nivel de fila é inversamente proporcional a prioridade) */
 ptFila fProx(){
-	/*
-	int i =  indexAtual+1;
-	if (i>=NFILAS-1)i = NFILAS-1;
-	return filas[i];
-	*/
-	return f1;
+	
+	//int i;// =  (indexAtual+1)%3;
+	//return filas[i];
+	switch (indexAtual)
+{
+   case 0:
+   	//indexAtual =1;
+     return f2;
+   break;
+   
+   case 1:
+     //indexAtual =2;
+     return f3;
+   break;
+      
+   case 2:
+    // indexAtual =2;
+     return f3;
+   break;
+   default: //erro???
+     return NULL;
+}
+	//indexAtual=i; //aqui que sera o controle?
+	return NULL;
+	
+	
+	//return f1;
 }
 
 /* Retorna fila, um nivel de fila menor do que a atual (note que o nivel de fila é inversamente proporcional a prioridade) */
 ptFila fAnt(){
-	/*
-	int i = indexAtual-1;
-	if (i<0)i = 0;
-	return filas[i];
-	*/
-	return f1;
+	
+	int i; //= indexAtual-1;
+	//if (i<0)i = NFILAS-1;
+	printf("FAnt index = %d\n",indexAtual);
+	switch (indexAtual)
+{
+   case 0:
+    //indexAtual=0;
+   	return f1;
+
+   case 1:
+   // indexAtual=0;
+   	return f1;
+      
+   case 2:
+  //  indexAtual=1;
+   	return f2;
+   default: //erro???
+     break;
+}
+	//indexAtual=i; //aqui que sera o controle?
+	
+	//return filas[i];
+	
+	return NULL;
 }
 
 void encerra(int status){
